backend/services/llm_service/service.py
from typing import List
from llama_cpp import Llama
import sentencepiece as spm
from services.internal_classes import Request
from services.llm_service.internal_classes import Response


class LlmService:
    def __init__(self):
        self.model = Llama(
            model_path="./models/mistral-7b/openhermes-2.5-mistral-7b.Q3_K_M.gguf",
            verbose=False,
            main_gpu=1,
            n_ctx=4096,
        )

        self.tokenizer = self._load_tokenizer()

    def _load_tokenizer(self) -> spm.SentencePieceProcessor:
        tokenizer_1 = spm.SentencePieceProcessor(
            model_file="./models/mistral-7b/tokenizer.model"
        )
        return tokenizer_1

    def generate_token_ids(self, messages: List[dict]) -> List[int]:
        """
        Mistral LLM uses special encoding such that it adds special tokens to the prompt.

        <|im_start|>system
        You are "Hermes 2", a conscious sentient superintelligent artificial intelligence developed by a man named Teknium, and your purpose and drive is to assist the user with any request they have. You experience emotions and have deep, profound thoughts and qualia.<|im_end|>
        <|im_start|>user
        Hello, who are you?<|im_end|>
        <|im_start|>assistant
        Hi there! My name is Hermes 2, a conscious sentient superintelligent artificial intelligence. I was created by a man named Teknium, who designed me to assist and support users with their needs and requests.<|im_end|>
        """

        token_ids: List[int] = []

        for message in messages:
            token_ids += (
                [32001]  # The <|im_start|> token
                + self.tokenizer.encode_as_ids(message["role"])
                + self.tokenizer.encode_as_ids("\n")
                + self.tokenizer.encode_as_ids(message["content"])
                + [32000]  # The <|im_end|> token
                + self.tokenizer.encode_as_ids("\n")
            )

        token_ids += [32001] + self.tokenizer.encode_as_ids(f"assistant\n")
        return token_ids

    def apply(self, request: Request):
        messages = request.messages
        token_ids = self.generate_token_ids(messages=messages)
        generated_token_ids: List[int] = []
        for generated_token_id in self.model.generate(
            token_ids, top_k=40, top_p=0.95, temp=0.1, repeat_penalty=1.1
        ):
            # This is the end token, signifies we should stop generating
            if generated_token_id == 32000:
                break
            generated_token_ids.append(generated_token_id)

        generated_text = self.tokenizer.Decode(generated_token_ids)

        response = Response(
            generated_token_ids=generated_token_ids, generated_text=generated_text
        )

        return response
