backend/requirements.txt
accelerate==0.25.0
annotated-types==0.6.0
anyio==3.7.1
beautifulsoup4==4.12.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
diskcache==5.6.3
fastapi==0.104.1
filelock==3.13.1
fsspec==2023.12.0
h11==0.14.0
huggingface-hub==0.19.4
idna==3.6
Jinja2==3.1.2
llama_cpp_python==0.2.20
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
protobuf==3.20.1
psutil==5.9.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
rsa==4.9
safetensors==0.4.1
sentencepiece==0.1.99
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
sympy==1.12
tokenizers==0.15.0
torch
torchvision
tqdm==4.66.1
transformers==4.35.2
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
